more fixes :(

+++ src/main/java/com/mo/mediaodyssey/models/DTO/CommunityMemberDTO.java
@@ -4,31 +4,20 @@ import com.mo.mediaodyssey.enums.RoleType;
 
 public class CommunityMemberDTO {
 
-    private final Integer userId;
+    private final Long userId;
     private final String username;
     private final RoleType roleType;
     private final String communityName;
 
-    public CommunityMemberDTO(Integer userId, String username, RoleType roleType, String communityName) {
+    public CommunityMemberDTO(Long userId, String username, RoleType roleType, String communityName) {
         this.userId = userId;
         this.username = username;
         this.roleType = roleType;
         this.communityName = communityName;
     }
 
-    public Integer getUserId() {
-        return userId;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public RoleType getRoleType() {
-        return roleType;
-    }
-
-    public String getCommunityName() {
-        return communityName;
-    }
+    public Long getUserId() { return userId; }
+    public String getUsername() { return username; }
+    public RoleType getRoleType() { return roleType; }
+    public String getCommunityName() { return communityName; }
 }
